Organizing includes and src in folders for makefile.

+++ src/player.cpp
@@ -49,12 +49,13 @@ void Player::moveLeft(Maze &maze, std::vector<Puzzle> &puzzles)
         {
             canMove = true;
         }
-        for(int i = 0; i < puzzles.size(); i++)
+        for(int i = 0; i < (int) puzzles.size(); i++)
             if(puzzles[i].isSolved(x, y - 1))
             {
                 puzzles.erase(puzzles.begin() + i);
                 canMove = true;
             }
+            
         if(x == endPos.first && y -  1 == endPos.second)
             canMove = true;
     }
@@ -66,18 +67,19 @@ void Player::moveLeft(Maze &maze, std::vector<Puzzle> &puzzles)
 void Player::moveRight(Maze &maze, std::vector<Puzzle> &puzzles)
 {
     bool canMove = false;
-    if(y < maze.getDims().second - 2 )
+    if(y < maze.getDims().second - 2)
     {
         if(maze.getTable()[x][y + 1] == ' ')
         {
             canMove = true;
         }
-        for(int i = 0; i < puzzles.size(); i++)
+        for(int i = 0; i < (int) puzzles.size(); i++)
             if(puzzles[i].isSolved(x, y + 1))
             {
                 puzzles.erase(puzzles.begin() + i);
                 canMove = true;
             }
+
         if(x == endPos.first && y + 1 == endPos.second)
             canMove = true;
     }
@@ -89,20 +91,21 @@ void Player::moveRight(Maze &maze, std::vector<Puzzle> &puzzles)
 void Player::moveUp(Maze &maze, std::vector<Puzzle> &puzzles)
 {
     bool canMove = false;
-    if( x > 1 )
+    if(x > 1)
     {
         if(maze.getTable()[x - 1][y] == ' ')
         {
             canMove = true;
         }
-        for(int i = 0; i < puzzles.size(); i++)
+        for(int i = 0; i < (int) puzzles.size(); i++)
             if(puzzles[i].isSolved(x - 1, y))
             {
                 puzzles.erase(puzzles.begin() + i);
                 canMove = true;
             }
-            if(x - 1 == endPos.first && y == endPos.second)
-                canMove = true;
+
+        if(x - 1 == endPos.first && y == endPos.second)
+            canMove = true;
         
     }
 
@@ -114,18 +117,19 @@ void Player::moveUp(Maze &maze, std::vector<Puzzle> &puzzles)
 void Player::moveDown(Maze &maze, std::vector<Puzzle> &puzzles)
 {
     bool canMove = false;
-    if( x < maze.getDims().first - 2 )
+    if(x < maze.getDims().first - 2)
     {
         if(maze.getTable()[x + 1][y] == ' ')
         {
             canMove = true;
         }
-        for(int i = 0; i < puzzles.size(); i++)
+        for(int i = 0; i < (int) puzzles.size(); i++)
             if(puzzles[i].isSolved(x + 1, y))
             {
                 puzzles.erase(puzzles.begin() + 1);
                 canMove = true;
             }
+
         if(x + 1 == endPos.first && y == endPos.second)
             canMove = true;
